requirements.txt
google-genai>=0.1.0
pillow>=9.0.0
python-dotenv>=0.19.0 